mg-platform/requirements.txt
# Dépendances du serveur MCP
mcp>=1.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.7.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.1.1

# Scrapers
scrapy>=2.12.0
playwright>=1.52.0

# Modèles de prévision
prophet>=1.1.0
pytorch-forecasting>=1.3.0
torch>=2.0.0
lightning>=2.0.0

# Tests
pytest>=8.3.0
